ring: alloc_ring_frames_generic make types size_t
Lets make i and num as size_t, there's no particular reason for them
to be int. At least i is used to setup iov_base offsets.

+++ ring.h
@@ -200,7 +200,7 @@ static inline int get_sockopt_tpacket(int sock)
 extern void setup_ring_layout_generic(int sock, struct ring *ring, size_t size,
 				      bool jumbo_support);
 extern void mmap_ring_generic(int sock, struct ring *ring);
-extern void alloc_ring_frames_generic(struct ring *ring, int num, size_t size);
+extern void alloc_ring_frames_generic(struct ring *ring, size_t num, size_t size);
 extern void bind_ring_generic(int sock, struct ring *ring, int ifindex, bool tx_only);
 
 #endif /* RING_H */
